Move Http\Client into Http package.
Move Cake\Network\Http\Client into the Http package. This change just
moves the implementation and not the test cases. This shows that the
correct class renames are in place. I wanted to move the implementation
first to keep this change small and easy to review. I've renamed
FormData\Part to FormDataPart as the extra directories were not
necessary.

I also wanted get consensus on the renames before doing more work.
