+++ src/Http/Client/Auth/Basic.php
@@ -11,7 +11,7 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Network\Http\Auth;
+namespace Cake\Http\Client\Auth;
 
 use Cake\Network\Http\Request;
