+++ src/Http/Client/Auth/Oauth.php
@@ -11,10 +11,10 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Network\Http\Auth;
+namespace Cake\Http\Client\Auth;
 
 use Cake\Core\Exception\Exception;
-use Cake\Network\Http\Request;
+use Cake\Http\Client\Request;
 
 /**
  * Oauth 1 authentication strategy for Cake\Network\Http\Client
