+++ src/Http/Client/CookieCollection.php
@@ -11,7 +11,9 @@
  * @since         3.0.0
  * @license       http://www.opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Network\Http;
+namespace Cake\Http\Client;
+
+use Cake\Http\Client\Response;
 
 /**
  * Container class for cookies used in Http\Client.
